Italian translation of config/locales/en.yml
This translation is based on the version relative to the last
config/locales/en.yml commit 917a95457f8e25826f

NOTE1: The accented characters have been written with HTML entities.

NOTE2: In the "number" section" (after line 114) the "separator",
"delimiter" and "unit" have been left as they were originally. Consider that
in Italy, the decimal separator is "," the thousands separator is "."
and the currency is Euro (Euro symbol).